Refactor classes to their appropriate components
