+++ src/main/java/seedu/address/logic/commands/TagCommand.java
@@ -16,8 +16,8 @@ import seedu.address.model.person.ContactIndex;
 import seedu.address.model.person.ModuleTagSet;
 import seedu.address.model.person.Person;
 import seedu.address.model.person.User;
-import seedu.address.model.scheduler.Module;
-import seedu.address.model.scheduler.time.Day;
+import seedu.address.model.timetable.Module;
+import seedu.address.model.time.Day;
 import seedu.address.model.tag.ModuleTag;
 
 /**
