+++ src/main/java/seedu/address/model/commitment/Lesson.java
@@ -5,9 +5,9 @@ import java.util.ArrayList;
 import org.joda.time.LocalTime;
 
 import seedu.address.model.location.Location;
-import seedu.address.model.scheduler.Module;
-import seedu.address.model.scheduler.time.Day;
-import seedu.address.model.scheduler.time.HourBlock;
+import seedu.address.model.timetable.Module;
+import seedu.address.model.time.Day;
+import seedu.address.model.time.HourBlock;
