+++ src/main/java/seedu/address/model/time/HourBlock.java
@@ -1,4 +1,4 @@
-package seedu.address.model.scheduler.time;
+package seedu.address.model.time;
 
 import java.util.Optional;
 
@@ -7,8 +7,8 @@ import org.joda.time.LocalTime;
 
 import seedu.address.model.commitment.Commitment;
 import seedu.address.model.scheduler.exceptions.CommitmentClashException;
-import seedu.address.model.scheduler.time.exceptions.WrongTimeException;
-import seedu.address.model.scheduler.time.util.TimeUtil;
+import seedu.address.model.time.exceptions.WrongTimeException;
+import seedu.address.model.time.util.TimeUtil;
 
 
 /**
