+++ src/main/java/seedu/address/model/time/TimeBlock.java
@@ -1,10 +1,10 @@
-package seedu.address.model.scheduler.time;
+package seedu.address.model.time;
 
 import org.joda.time.Hours;
 import org.joda.time.LocalTime;
 
-import seedu.address.model.scheduler.time.exceptions.WrongTimeException;
-import seedu.address.model.scheduler.time.util.TimeUtil;
+import seedu.address.model.time.exceptions.WrongTimeException;
+import seedu.address.model.time.util.TimeUtil;
 
 /**
  * Represents a time period of a variable amount.
