+++ src/main/java/seedu/address/model/time/exceptions/WrongTimeException.java
@@ -1,4 +1,4 @@
-package seedu.address.model.scheduler.time.exceptions;
+package seedu.address.model.time.exceptions;
 
 /**
  * Represents a timing error.
