+++ src/main/java/seedu/address/model/time/util/TimeUtil.java
@@ -1,4 +1,4 @@
-package seedu.address.model.scheduler.time.util;
+package seedu.address.model.time.util;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -8,12 +8,12 @@ import java.util.stream.Stream;
 import org.joda.time.LocalTime;
 
 import javafx.util.Pair;
+import seedu.address.model.time.Day;
+import seedu.address.model.time.TimeBlock;
+import seedu.address.model.time.TimePeriod;
 import seedu.address.model.timetable.Timetable;
-import seedu.address.model.scheduler.time.Day;
-import seedu.address.model.scheduler.time.HourBlock;
-import seedu.address.model.scheduler.time.TimeBlock;
-import seedu.address.model.scheduler.time.TimePeriod;
-import seedu.address.model.util.MathUtil;
+import seedu.address.model.time.HourBlock;
+import seedu.address.commons.util.MathUtil;
 
 
 /**
