+++ src/test/java/seedu/address/model/timetable/LessonTest.java
@@ -16,8 +16,7 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.model.commitment.Lesson;
 import seedu.address.model.location.Location;
-import seedu.address.model.scheduler.Module;
-import seedu.address.model.scheduler.time.Day;
+import seedu.address.model.time.Day;
 
 class LessonTest {
