+++ src/test/java/seedu/address/model/timetable/ModuleTest.java
@@ -14,7 +14,6 @@ import java.util.List;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.model.commitment.Lesson;
-import seedu.address.model.scheduler.Module;
 
 class ModuleTest {
     @Test
