+++ src/test/java/seedu/address/model/timetable/time/HourBlockTest.java
@@ -17,11 +17,11 @@ import org.joda.time.LocalTime;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.model.scheduler.exceptions.CommitmentClashException;
-import seedu.address.model.scheduler.time.Day;
-import seedu.address.model.scheduler.time.HourBlock;
-import seedu.address.model.scheduler.time.TimeBlock;
-import seedu.address.model.scheduler.time.TimePeriod;
-import seedu.address.model.scheduler.time.exceptions.WrongTimeException;
+import seedu.address.model.time.Day;
+import seedu.address.model.time.HourBlock;
+import seedu.address.model.time.TimeBlock;
+import seedu.address.model.time.TimePeriod;
+import seedu.address.model.time.exceptions.WrongTimeException;
 
 class HourBlockTest {
