+++ src/test/java/seedu/address/model/timetable/time/TimetableTest.java
@@ -36,9 +36,9 @@ import java.util.stream.IntStream;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.model.scheduler.Timetable;
+import seedu.address.model.timetable.Timetable;
 import seedu.address.model.scheduler.exceptions.CommitmentClashException;
-import seedu.address.model.scheduler.time.HourBlock;
+import seedu.address.model.time.HourBlock;
 
 class TimetableTest {
