+++ src/test/java/seedu/address/model/timetable/util/TypicalLesson.java
@@ -4,8 +4,8 @@ import org.joda.time.LocalTime;
 
 import seedu.address.model.commitment.Lesson;
 import seedu.address.model.location.Location;
-import seedu.address.model.scheduler.Module;
-import seedu.address.model.scheduler.time.Day;
+import seedu.address.model.timetable.Module;
+import seedu.address.model.time.Day;
 
 
 /**
